feat/Tests (#14)
* chore: format Expr.g4

* fix: generate output in paste bin

* fix: trying to fix, scanf, structs and unions

* fix: scanf, structs, unions and boolean

* fix: structs, puts, gets, boolean

* fix: string conversions

* fix: expr order, bool, string and escape sequence

* fix: Expr, increment, include and compound

* fix: EscapeSequenceProcessor, comparison and type convertions

* test: creating tests for control flow

* test: loops

* test: functions

* test: pointers and dynamic memory

* test: structs and unions

* fix: structs and unions

* tests: integration and language_tests

* feat: unit tests for semantic analyzers

* fix: DynamicMemoryCodeGenerator

* chore: fix tests, puts, and update readme

+++ Three_Musketeers/Program.cs
@@ -11,7 +11,7 @@ namespace Three_Musketeers
 {
     public class Program
     {
-        public static readonly string VERSION = "v0.9.5";
+        public static readonly string VERSION = "v1.0.0 Athos";
         private static readonly CliArgument<string> input = new("input") {
             Description = "Path of the source file to compile",
             Arity = ArgumentArity.ZeroOrOne
@@ -61,7 +61,8 @@ namespace Three_Musketeers
                 compilerOptimazionLevel,
                 dotLLCodePath,
                 addDebugFlagToGcc,
-                includeLibrary
+                includeLibrary,
+                generateBin
             }!;
 
             foreach (var option in rootCommand.Options)
@@ -180,8 +181,24 @@ namespace Three_Musketeers
                 string executablePath = Path.Combine(result, resultPath);
                 File.WriteAllText(outputPath, llvmCode);
 
-                Process.Start("llc", $"{outputPath} -O{optimizationLevel} -o {assemblyFilePath}")?.WaitForExit();
-                Process.Start("gcc", $"{(addDebugFlag ? "-g" : "")} {assemblyFilePath} -o {executablePath} -no-pie")?.WaitForExit();
+                var llcProcess = Process.Start("llc", $"{outputPath} -O{optimizationLevel} -o {assemblyFilePath}");
+                llcProcess?.WaitForExit();
+
+                if (llcProcess?.ExitCode != 0)
+                {
+                    WriteError($"LLVM Compilation failed (llc exited with code {llcProcess?.ExitCode})");
+                    return 1;
+                }
+
+                var gccProcess = Process.Start("gcc", $"{(addDebugFlag ? "-g" : "")} {assemblyFilePath} -o {executablePath} -no-pie");
+                gccProcess?.WaitForExit();
+
+                if (gccProcess?.ExitCode != 0)
+                {
+                    WriteError($"GCC Linking failed (gcc exited with code {gccProcess?.ExitCode})");
+                    return 1;
+                }
+
                 File.Delete(assemblyFilePath);
 
                 if (!saveLLVM)
